db: the first five queries are checked at compile time
All 67 queries were runtime strings: a typo'd column, a wrong type or a
dropped table compiled cleanly and failed when the path was first hit in
production. sqlx's macros verify SQL against the real schema at build time.
Proved by breaking one on purpose — 'Unknown column bytez' now stops the
build, naming the line.

It found a type lie immediately. difficulty is tinyint(4), while the row
struct and the API both called it i32; the runtime path was silently
widening it. The row type says i8 now and widens on the way out, so the
public shape is unchanged.

.sqlx is the offline cache that lets this build with no database, in CI and
in the nix sandbox — verified by compiling with DATABASE_URL unset. An
edited query fails on its own, its text being the cache key, but a MIGRATION
can age the cache while every query still matches and builds green. Hence a
gate row that applies the migrations to an ephemeral server and asks sqlx
whether the cache still describes it. Checked both ways: green as committed,
and rc=1 naming the file when a cached column type is altered.

The remaining 62 queries still take the runtime path.

+++ src/exercise/types.rs
@@ -220,7 +220,10 @@ pub(crate) struct ExerciseDetailRow {
     pub cue: Option<String>,
     pub demo_url: Option<String>,
     pub summary: Option<String>,
-    pub difficulty: Option<i32>,
+    /// `tinyint(4)` in the schema, so `i8` here and widened on the way out. The
+    /// runtime query path silently widened it, which is what a checked query
+    /// stops: the row type now says what the column actually is.
+    pub difficulty: Option<i8>,
     pub has_image: i64,
     pub has_loop: i64,
 }
